requirements.txt
asgiref==3.9.2
bleach==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.6
django-ckeditor==6.7.3

django-js-asset==3.1.2
idna==3.10
pillow==11.3.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
gunicorn==21.2.0
psycopg[binary]>=3.2.0
dj-database-url>=2.1.0
python-dotenv>=1.0.1




